REF: share Index/Block get_values_for_csv
