+++ pandas/core/indexes/interval.py
@@ -847,7 +847,7 @@ class IntervalIndex(ExtensionIndex):
 
     def _format_with_header(self, *, header: list[str], na_rep: str) -> list[str]:
         # matches base class except for whitespace padding
-        return header + list(self._format_native_types(na_rep=na_rep))
+        return header + list(self._get_values_for_csv(na_rep=na_rep))
 
     # --------------------------------------------------------------------
     # Set Operations
