+++ pandas/tests/indexes/interval/test_formats.py
@@ -104,7 +104,7 @@ class TestIntervalIndexRendering:
     def test_to_native_types(self, tuples, closed, expected_data):
         # GH 28210
         index = IntervalIndex.from_tuples(tuples, closed=closed)
-        result = index._format_native_types(na_rep="NaN")
+        result = index._get_values_for_csv(na_rep="NaN")
         expected = np.array(expected_data)
         tm.assert_numpy_array_equal(result, expected)
